fix certain structures and slightly refactor CacheManager

+++ src/main/java/botcommons/utilities/JsonUtils.java
@@ -2,24 +2,17 @@ package botcommons.utilities;
 
 import com.google.gson.*;
 import com.google.gson.reflect.TypeToken;
-import com.google.gson.stream.JsonReader;
-import com.google.gson.stream.JsonWriter;
-import lombok.EqualsAndHashCode;
 import net.dv8tion.jda.api.JDA;
-import net.dv8tion.jda.api.entities.GuildVoiceState;
 import net.dv8tion.jda.api.entities.Member;
-import net.dv8tion.jda.internal.utils.IOUtil;
 
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.lang.ref.Reference;
 import java.lang.reflect.Modifier;
 import java.lang.reflect.Type;
 import java.nio.file.Path;
 import java.util.HashMap;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 public class JsonUtils {
 	private static final Gson GSON = new GsonBuilder()
@@ -42,6 +35,7 @@ public class JsonUtils {
 
 	public static <T, V> void createCache(HashMap<T, V> map, String name) {
 		File file = StringUtilities.getAssetPath(Path.of("cache/",name+".json")).toFile();
+		if (file.exists()) return;
 		try (FileWriter writer = new FileWriter(file)) {
 			GSON.toJson(map, writer);
 		} catch (IOException e) {
@@ -58,6 +52,7 @@ public class JsonUtils {
 			try (FileReader reader = new FileReader(file)) {
 				Type type = new TypeToken<HashMap<?, ?>>() {}.getType();
 				existingData.putAll(GSON.fromJson(reader, type));
+				System.out.println("Loaded existing data from " + file.getName());
 			} catch (IOException e) {
 				e.printStackTrace();
 			}
